Added 4 Cypress Tests for QA Automation Assessment
In this commit, I have introduced four comprehensive automated tests using Cypress to address key scenarios outlined in the QA Automation Assessment. These tests rigorously validate the functionality of the application and ensure its reliability.

- Locating and Editing Computer Details:

I have implemented a test to locate the computer named "Commodore 64" within the table.
Subsequently, I simulated a click to edit its details.
Furthermore, I designed a negative test case to trigger a failure validation message, confirming the robustness of the system.
Additionally, a positive test case was established to verify the successful update of valid data.

- Filtering and Mapping Data for "HP" Computers:

The second test involves filtering the list of computers by "HP."
I then created a comprehensive map of the data retrieved from this filter.
The generated map is thoroughly examined to ensure accurate representation.

- Extracting Computer Names for "IBM" on the Last Page:

Another critical test was developed to filter the computer list by "IBM."
The focus of this test is to extract the names of computers situated on the last page of the search results.
The extracted list of computer names is meticulously verified to ensure correctness.

- Creating a New Computer Entry:

This test case is centered around the process of adding a new computer entry.
I meticulously simulated clicking the 'Add a new computer' button, populating various text fields.
I also selected "Evans & Sutherland" from the company dropdown.
Subsequently, I confirmed the creation of the computer entry and verified its success, ensuring the seamless functionality of this essential feature.